requirements.txt
pyrogram
python-dotenv
motor
aiofiles